pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain-anthropic>=1.4.0",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.6",
    "langsmith>=0.7.32",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "tavily-python>=0.7.23",
]
